+++ projects/fhi-angular-components/src/lib/fhi-time/fhi-month.model.ts
@@ -1,0 +1,4 @@
+export class FhiMonth {
+  year: number;
+  month: number;
+}