+++ projects/fhi-angular-components/src/lib/fhi-time/fhi-year-month-range/fhi-month-range.component copy.ts
@@ -1,34 +1,28 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { isValid } from 'date-fns';
-
 import { FhiAutosuggestModule } from '../../fhi-autosuggest/fhi-autosuggest.module';
 import { FhiAutosuggestItem } from '../../fhi-autosuggest/fhi-autosuggest.model';
-import { FhiYearSelectorComponent } from '../../fhi-year-selector/fhi-year-selector.component';
+import { FhiYearsComponent } from '../fhi-years/fhi-years.component';
 import { FhiConstantsService } from '../../shared-services/fhi-constants.service';
 
 @Component({
   selector: 'fhi-month-range',
   standalone: true,
-  imports: [
-    CommonModule,
-    FhiAutosuggestModule,
-    FhiYearSelectorComponent
-  ],
-  templateUrl: './fhi-month-range.component.html',
-  providers: [ FhiConstantsService ]
+  imports: [CommonModule, FhiAutosuggestModule, FhiYearsComponent],
+  templateUrl: './fhi-month-range-copy.component.html',
+  providers: [FhiConstantsService],
 })
-export class FhiMonthRangeComponent {
+export class FhiMonthRangeCopyComponent {
   @Input() maxYear: number = this.FHI_CONSTANTS.MAX_YEAR;
   @Input() minYear: number = this.FHI_CONSTANTS.MIN_YEAR;
-  
-  @Output() monthRangeSelect = new EventEmitter<Object>();
-  
-  fieldsetLegendFrom: string = 'Fra måned';
-  fieldsetLegendTo: string = 'Til måned';
-  labelMonth: string = 'måned';
-  labelYear: string = 'år';
+
+  @Output() monthRangeSelect = new EventEmitter<object>();
+
+  fieldsetLegendFrom = 'Fra måned';
+  fieldsetLegendTo = 'Til måned';
+  labelMonth = 'måned';
+  labelYear = 'år';
   maxYearFrom: number;
   minYearTo: number;
   monthListFull: FhiAutosuggestItem[] = [
@@ -43,13 +37,13 @@ export class FhiMonthRangeComponent {
     { id: 9, name: 'September' },
     { id: 10, name: 'Oktober' },
     { id: 11, name: 'November' },
-    { id: 12, name: 'Desember'}
+    { id: 12, name: 'Desember' },
   ];
   monthFrom: number;
   monthFromList: FhiAutosuggestItem[] = [...this.monthListFull];
   monthTo: number;
   monthToList: FhiAutosuggestItem[] = [...this.monthListFull];
-  validRange: boolean = true;
+  validRange = true;
   yearFrom: string;
   yearTo: string;
 
@@ -59,6 +53,11 @@ export class FhiMonthRangeComponent {
     this.maxYearFrom = this.maxYear;
     this.minYearTo = this.minYear;
   }
+
+  // onYearSelect(year: number[], context: string) {
+  //   console.log('year', year);
+  //   console.log('context', context);
+  // }
 
   onYearFromSelect(event: any) {
     this.yearFrom = event;
@@ -72,6 +71,11 @@ export class FhiMonthRangeComponent {
     this.checkValidity();
   }
 
+  // onMonthSelect(month: number[], context: string) {
+  //   console.log('month', month);
+  //   console.log('context', context);
+  // }
+
   onMonthFromSelect(event: any) {
     this.monthFrom = event;
     this.checkValidity();
@@ -83,31 +87,31 @@ export class FhiMonthRangeComponent {
   }
 
   private checkValidity() {
-    const fromMonth: any = new Date(Number(this.yearFrom), this.monthFrom -1);
+    const fromMonth: any = new Date(Number(this.yearFrom), this.monthFrom - 1);
     const toMonth: any = new Date(Number(this.yearTo), this.monthTo, 0);
-    
+
     if (toMonth - fromMonth >= 0) {
       this.monthRangeSelect.emit({
         start: {
           year: Number(this.yearFrom),
-          month: this.monthFrom
+          month: this.monthFrom,
         },
         end: {
           year: Number(this.yearTo),
-          month: this.monthTo
-        }
+          month: this.monthTo,
+        },
       });
       this.validRange = true;
     }
     if (toMonth - fromMonth < 0) {
       this.validRange = false;
     }
-    
+
     if (
-        this.yearFrom === this.yearTo &&
-        this.yearFrom !== undefined &&
-        this.yearTo !== undefined
-      ) {
+      this.yearFrom === this.yearTo &&
+      this.yearFrom !== undefined &&
+      this.yearTo !== undefined
+    ) {
       this.adjustMonthList();
     } else {
       this.monthFromList = [...this.monthListFull];